+++ package.json
@@ -23,6 +23,7 @@
   "author": "tascript",
   "license": "MIT",
   "dependencies": {
+    "@actions/core": "^1.6.0",
     "chalk": "^4.1.2"
   },
   "devDependencies": {
